gh-320  extract into files
